+++ rubato/utils/rb_time.py
@@ -10,10 +10,22 @@ from . import InitError
 
 @dataclass(order=True)
 class DelayedTask:
-    """A task that is run after a specified number of milliseconds."""
+    """
+    A task that is run after a specified number of milliseconds.
+
+    Args:
+        delay: The number of milliseconds to wait before invoking the task.
+        task: The task to invoke.
+        is_stopped: Whether the task is stopped.
+    """
     delay: int
+    """The delay until the task is run, in milliseconds."""
     task: Callable[[], None] = field(compare=False)
+    """The task to run."""
     is_stopped: bool = field(default=False, compare=False)
+    """Whether the DelayedTask is stopped."""
+    next_run: int = field(default=0, compare=False)
+    """The time at which the task will be run next, in milliseconds."""
 
     def stop(self):
         """Stop the DelayedTask from invoking."""
@@ -22,10 +34,22 @@ class DelayedTask:
 
 @dataclass(order=True)
 class FramesTask:
-    """A task that is run after a specified number of frames."""
+    """
+    A task that is run after a specified number of frames.
+
+    Args:
+        delay: The number of frames to wait before invoking the task.
+        task: The task to invoke.
+        is_stopped: Whether the task is stopped.
+    """
     delay: int
+    """The delay until the task is run, in frames."""
     task: Callable[[], None] = field(compare=False)
+    """The task to run."""
     is_stopped: bool = field(default=False, compare=False)
+    """Whether the FramesTask is stopped."""
+    next_run: int = field(default=0, compare=False)
+    """The frame at which the task will be run next."""
 
     def stop(self):
         """Stop the FramesTask from invoking."""
@@ -33,15 +57,29 @@ class FramesTask:
 
 
 @dataclass(order=True)
-class ScheduledTask:
-    """A task that is run every specified number of milliseconds."""
+class RecurrentTask:
+    """
+    A task that is run every specified number of milliseconds.
+
+    Args:
+        interval: The number of milliseconds between task invocations.
+        task: The task to invoke.
+        delay: The number of milliseconds to wait before starting the invocations.
+        is_stopped: Whether the task is stopped.
+    """
     interval: int = field(compare=False)
-    task: Callable[[], None] | Callable[["ScheduledTask"], None] = field(compare=False)
+    """The interval between task invocations, in milliseconds."""
+    task: Callable[[], None] | Callable[["RecurrentTask"], None] = field(compare=False)
+    """The task to run."""
     delay: int = field(default=0)
+    """The initial delay until the task is run, in milliseconds."""
     is_stopped: bool = field(default=False, compare=False)
+    """Whether the RecurrentTask is stopped."""
+    next_run: int = field(default=0, compare=False)
+    """The time at which the task will be run next, in milliseconds."""
 
     def stop(self):
-        """Stop the ScheduledTask from invoking."""
+        """Stop the RecurrentTask from invoking."""
         self.is_stopped = True
 
 
@@ -56,7 +94,7 @@ class Time:
 
     _frame_queue: list[FramesTask] = []
     _task_queue: list[DelayedTask] = []
-    _schedule_queue: list[ScheduledTask] = []
+    _recurrent_queue: list[RecurrentTask] = []
 
     _next_queue: list[Callable] = []
 
@@ -145,7 +183,7 @@ class Time:
             func: The function to call
         """
 
-        heapq.heappush(cls._frame_queue, FramesTask(cls.frames + delay, func))
+        cls.schedule(FramesTask(delay, func))
 
     @classmethod
     def delayed_call(cls, delay: int, func: Callable[[], None]):
@@ -157,39 +195,40 @@ class Time:
             func: The function to call.
         """
 
-        heapq.heappush(cls._task_queue, DelayedTask(delay + cls.now(), func))
+        cls.schedule(DelayedTask(delay, func))
 
     @classmethod
-    def scheduled_call(cls, interval: int, func: Callable[[], None] | Callable[[ScheduledTask], None]):
+    def recurred_call(cls, interval: int, func: Callable[[], None] | Callable[[RecurrentTask], None], delay: int = 0):
         """
         Schedules the function func to be repeatedly called every interval.
 
         Args:
             interval: The interval (in milliseconds) to run the function at.
-            func: The function to call. Can take a ScheduledTask as an argument. This is useful for stopping the task.
+            func: The function to call. Can take a RecurrentTask as an argument. This is useful for stopping the task.
+            delay: The delay (in milliseconds) to wait before starting the task.
         """
 
-        heapq.heappush(cls._schedule_queue, ScheduledTask(interval, func, interval + cls.now()))
+        cls.schedule(RecurrentTask(interval, func, delay))
 
     @classmethod
-    def schedule(cls, task: DelayedTask | FramesTask | ScheduledTask):
+    def schedule(cls, task: DelayedTask | FramesTask | RecurrentTask):
         """
         Schedules a task for execution based on what type of task it is.
 
         Args:
-            task (DelayedTask | FramesTask | ScheduledTask): The task to queue.
+            task: The task to queue.
         """
         if isinstance(task, DelayedTask):
-            task.delay += cls.now()
+            task.next_run = cls.now() + task.delay if task.next_run <= cls.now() else task.next_run
             heapq.heappush(cls._task_queue, task)
         elif isinstance(task, FramesTask):
-            task.delay += cls.frames
+            task.next_run = cls.frames + task.delay if task.next_run <= cls.frames else task.next_run
             heapq.heappush(cls._frame_queue, task)
-        elif isinstance(task, ScheduledTask):
-            task.delay += cls.now()
-            heapq.heappush(cls._schedule_queue, task)
+        elif isinstance(task, RecurrentTask):
+            task.next_run = cls.now() + task.delay + task.interval if task.next_run <= cls.now() else task.next_run
+            heapq.heappush(cls._recurrent_queue, task)
         else:
-            raise TypeError("Task argument must of of type DelayedTask, FramesTask or ScheduledTask.")
+            raise TypeError("Task argument must of of type DelayedTask, FramesTask or RecurrentTask.")
 
     @classmethod
     def milli_to_sec(cls, milli: float | int) -> float | int:
@@ -226,33 +265,32 @@ class Time:
             cls._next_queue.clear()
 
         while cls._frame_queue:
-            if cls._frame_queue[0].delay <= cls.frames:
-                timer_task = heapq.heappop(cls._frame_queue)
-                if not timer_task.is_stopped:
-                    timer_task.task()
+            if cls._frame_queue[0].next_run <= cls.frames:
+                frame_task: FramesTask = heapq.heappop(cls._frame_queue)
+                if not frame_task.is_stopped:
+                    frame_task.task()
             else:
                 break
 
         while cls._task_queue:
-            if cls._task_queue[0].delay <= cls.now():
-                timer_task = heapq.heappop(cls._task_queue)
-                if not timer_task.is_stopped:
-                    timer_task.task()
+            if cls._task_queue[0].next_run <= cls.now():
+                task: DelayedTask = heapq.heappop(cls._task_queue)
+                if not task.is_stopped:
+                    task.task()
             else:
                 break
 
-        while cls._schedule_queue:
-            if cls._schedule_queue[0].delay <= cls.now():
-                scheduled_task = heapq.heappop(cls._schedule_queue)
+        while cls._recurrent_queue:
+            if cls._recurrent_queue[0].next_run <= cls.now():
+                recurrent_task: RecurrentTask = heapq.heappop(cls._recurrent_queue)
 
-                if not scheduled_task.is_stopped:
+                if not recurrent_task.is_stopped:
                     try:
-                        scheduled_task.task(scheduled_task)  # type: ignore
+                        recurrent_task.task(recurrent_task)  # type: ignore
                     except TypeError:
-                        scheduled_task.task()  # type: ignore
+                        recurrent_task.task()  # type: ignore
 
-                if not scheduled_task.is_stopped:
-                    scheduled_task.delay += scheduled_task.interval
-                    heapq.heappush(cls._schedule_queue, scheduled_task)
+                    recurrent_task.next_run += recurrent_task.interval
+                    heapq.heappush(cls._recurrent_queue, recurrent_task)
             else:
                 break